+++ dessia_common/serialization.py
@@ -28,6 +28,7 @@ fullargsspec_cache = {}
 class SerializableObject(CoreDessiaObject):
     """ Serialization capabilities of Dessia Object. """
 
+    _standalone_in_db = False
     _non_serializable_attributes = []
 
     def base_dict(self):
@@ -67,7 +68,7 @@ class SerializableObject(CoreDessiaObject):
         serialized_dict = self.base_dict()
         dict_ = self._serializable_dict()
         if use_pointers:
-            serialized_dict.update(serialize_dict_with_pointers(dict_, memo, path, id_method=id_method,
+            serialized_dict.update(serialize_dict_with_pointers(dict_, memo=memo, path=path, id_method=id_method,
                                                                 id_memo=id_memo)[0])
         else:
             serialized_dict.update(serialize_dict(dict_))
@@ -149,25 +150,32 @@ def serialize_with_pointers(value, memo=None, path='#', id_method=True, id_memo=
         memo = {}
     if id_memo is None:
         id_memo = {}
+
     if isinstance(value, SerializableObject):
         if value in memo:
-            return {'$ref': memo[value]}, memo
+            path_value, serialized_value, id_ = memo[value]
+            id_memo[id_] = serialized_value
+            return {'$ref': path_value}, memo
         try:
             serialized = value.to_dict(use_pointers=True, memo=memo, path=path, id_memo=id_memo)
 
         except TypeError:
-            warnings.warn('specific to_dict should implement use_pointers, memo and path arguments', Warning)
+            warnings.warn('specific to_dict should implement use_pointers, memo, path and id_memo arguments', Warning)
             serialized = value.to_dict()
         if id_method:
             id_ = str(uuid.uuid1())
-            id_memo[id_] = serialized
-            memo[value] = f'#/_references/{id_}'
-            serialized = {'$ref': memo[value]}
+            path_value = f"#/_references/{id_}"
+            memo[value] = path_value, serialized, id_
+            if value._standalone_in_db:
+                id_memo[id_] = serialized
+                serialized = {'$ref': path_value}
         else:
-            memo[value] = path
+            memo[value] = path, serialized, None
 
     elif isinstance(value, type):
         if value in memo:
+            path_value, serialized_value, id_ = memo[value]
+            id_memo[id_] = serialized_value
             return {'$ref': memo[value]}, memo
         serialized = serialize_typing(value)
         # memo[value] = path
@@ -175,16 +183,17 @@ def serialize_with_pointers(value, memo=None, path='#', id_method=True, id_memo=
     # Regular object
     elif hasattr(value, 'to_dict'):
         if value in memo:
-            return {'$ref': memo[value]}, memo
+            path_value, serialized_value, id_ = memo[value]
+            id_memo[id_] = serialized_value
+            return {'$ref': path}, memo
         serialized = value.to_dict()
 
         if id_method:
             id_ = str(uuid.uuid1())
-            id_memo[id_] = serialized
-            memo[value] = f'#/_references/{id_}'
-            serialized = {'$ref': memo[value]}
+            path_value = f"#/_references/{id_}"
+            memo[value] = path_value, serialized, id_
         else:
-            memo[value] = path
+            memo[value] = path, serialized, None
 
     elif isinstance(value, dict):
         serialized, memo = serialize_dict_with_pointers(value, memo=memo, path=path, id_method=id_method,
@@ -203,7 +212,7 @@ def serialize_with_pointers(value, memo=None, path='#', id_method=True, id_memo=
             raise dc_err.SerializationError(msg)
         serialized = value
 
-    if path == '#':
+    if path == '#' and id_method:
         # adding _references
         serialized['_references'] = id_memo
 
@@ -425,7 +434,7 @@ def deserialize_with_typing(type_, argument, global_dict=None, pointers_memo=Non
             deserialized_arg = iter(deserialized_arg)
 
     elif origin is tuple:
-        # Heterogenous sequences (tuples)
+        # Heterogeneous sequences (tuples)
         deserialized_arg = tuple(deserialize_argument(t, arg) for (t, arg) in zip(args, argument))
     elif origin is dict:
         # Dynamic dict
@@ -613,7 +622,11 @@ def pointer_graph(value):
 
 
 def update_pointers_data(global_dict, current_dict, pointers_memo):
-    """ Update pointers according to cuccrent dict. """
+    """
+    Update pointers according to current dict.
+    
+    :returns: the global dict and a pointer memo
+    """
     if global_dict is None or pointers_memo is None:
         global_dict = current_dict
 
@@ -654,7 +667,7 @@ def deserialization_order(dict_):
 
 def dereference_jsonpointers(dict_):  # , global_dict):
     """
-    Analyse given dict.
+    Analyses given dict.
 
      Useful in order to:
     - find jsonpointers