+++ tests/ci_tests.py
@@ -41,6 +41,8 @@ scripts = [
     "workflow/various_tests.py",
     "workflow/workflow_building.py",
     "workflow/workflow_sampling.py",
+    "workflow/workflow_inputs.py",
+    "workflow/file_inputs.py",
 
     # Breakdown
     "breakdown.py",