+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.14.0 [now dev]
+
+### CI
+
+- fixing versions of pylint and pydocstyle to avoid uncontrolled new errors
 
 ## 0.14.0
 
@@ -37,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor search_memo function to improve copy performance
 - Add pre-checks for non list-like element in is_sequence function
 - For is_serializable, not using pointers while trying to_dict
+- Refactor workflow run values for serializability checks
 
 
 ### Chore
@@ -51,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Reference path is now given all the way down to plot_data
+- DessiaObject kwargs in init are deprecated
 
 ### Fixed